Base version-agnostic OSNodeDriver class.
Version-specific subclasses will inherit from this. I expect more will
be moved into it in coming commits.
