+++ demos/secrets.py-dist
@@ -24,8 +24,7 @@ ECP_PARAMS = ('user_name', 'password')
 GANDI_PARAMS = ('user',)
 HOSTINGCOM_PARAMS = ('user', 'secret')
 IBM_PARAMS = ('user', 'secret')
-# OPENSTACK_PARAMS = ('user_name', 'api_key', secure_bool, 'host', port_int)
-OPENSTACK_PARAMS = ('user_name', 'api_key', False, 'host', 8774)
+OPENSTACK_PARAMS = ('user_name', 'api_key', 'auth_url')
 OPENNEBULA_PARAMS = ('user', 'key')
 OPSOURCE_PARAMS = ('user', 'password')
 RACKSPACE_PARAMS = ('user', 'key')
